LayoutTests/accessibility/mac/selected-visible-position-range-expected.txt

First paragraph.

Last paragraph.

This tests getting and setting the selected VisiblePosition range.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS selectedString is 'Last paragraph.'
selectedString must be the text of the first paragraph: First paragraph.


selectedString must be an empty string: 
selectedString must be an empty string: 
PASS successfullyParsed is true

TEST COMPLETE
